// proto/book/BookResponse.ts
export interface BookResponse {
  'id'?: (string);
  'title'?: (string);
  'authorId'?: (string);
  'publishedYear'?: (number);
  'createdAt'?: (string);
}

export interface Book__Output {
  'id': (string);
  'title': (string);
  'authorId': (string);
  'publishedYear': (number);
  'createdAt': (string);
}